Error Solve

+++ application/modules/prescription/controllers/Prescription.php
@@ -499,7 +499,9 @@ class Prescription extends MX_Controller {
 
             $option1 = '<a title="' . lang('view') . ' ' . lang('prescription') . '" class="btn btn-info btn-xs btn_width" href="prescription/viewPrescription?id=' . $prescription->id . '"><i class="fa fa-eye"> ' . lang('view') . ' ' . lang('prescription') . ' </i></a>';
             $option3 = '<a class="btn btn-info btn-xs btn_width" href="prescription/editPrescription?id=' . $prescription->id . '" data-id="' . $prescription->id . '"><i class="fa fa-edit"></i> ' . lang('edit') . ' ' . lang('prescription') . '</a>';
+          if (!$this->ion_auth->in_group(array('Pharmacist', 'Receptionist')) || $permis_2 =='ok') {
             $option2 = '<a class="btn btn-info btn-xs btn_width delete_button" href="prescription/delete?id=' . $prescription->id . '&admin=' . $prescription->id . '" onclick="return confirm(\'Are you sure you want to delete this item?\');"><i class="fa fa-trash"> </i></a>';
+          }
             $options4 = '<a class="btn btn-info btn-xs invoicebutton" title="' . lang('print') . '" style="color: #fff;" href="prescription/viewPrescriptionPrint?id=' . $prescription->id . '"target="_blank"> <i class="fa fa-print"></i> ' . lang('print') . '</a>';
 
             if (!empty($prescription->medicine)) {
@@ -533,10 +535,10 @@ class Prescription extends MX_Controller {
                 $option3 = '';
                 
             }
-              if ($this->ion_auth->in_group(array('Pharmacist', 'Receptionist')) || $permis_2 !='ok') {
+           /* if ($this->ion_auth->in_group(array('Pharmacist', 'Receptionist')) || $permis_2 !='ok') {
                 $option2 = '';
                 
-            }
+            }*/
             $info[] = array(
                 $prescription->id,
                 date('d-m-Y', $prescription->date),
